Revert "Merge branch 'main' into mixing-rules"
This reverts commit c5dfc9e570ad178b4e8139baaeb52df35e46fb62, reversing
changes made to cae296db2a5b6427609b61b55af28666a5bf4fb0.

Revert non-working merge

+++ src/IO/input/SimParameters.h
@@ -71,7 +71,7 @@ public:
     std::string getStoreCheckpoint() const { return store_checkpoint_path ;}
     std::string getloadCheckpoint() const { return load_checkpoint_path ;}
     double getInitTemperature()const {return  initTemperature;}
-    int getThermostatCycleLength()const{return  thermostatCycleLength;}
+    double getThermostatCycleLength()const{return  thermostatCycleLength;}
     double getTargetTemperature()const{return targetTemperature;}
     double getMaxTemperatureChange()const{return maxTemperatureChange;}
